requirements.txt
django>=1.10,<1.11
django-rest-swagger
djangorestframework==3.6.4
